Prune auto-research legacy core
